First working version of A&D UA-651BLE sync
The code in /AND_UA651BLE/2_Sync continously listens for a specific Blood Pressure Monitor and connects to sync when it advertises.
It is also able to connect to a Adafruit Bluefruit LE UART Friend to read and write bytes through it's serial port.
Lastly it connects to an MQTT broker to publish blood pressure measurements, and subscribe to notifications to send to the serial-device.