Fixed pairing with AND UA-651BLE
The pairing application now starts by scanning for compatible devices
based on the vendor-part of their MAC address, and initiates bonding and
setup of encryption keys to discovered devices.

In addition, the datetime characteristic of the blood pressure service is
written to the current UTC+1 time based on the computer clock, to
synchronize the time of the device.